fix(a2-audit): Paso 2 - migrar a integer 18 casos en Catalog y Orders
FASE A2 - PASO 2: Migración de 143 casts (float) residuales.

CAMBIOS EN CÓDIGO (18 cambios en 6 archivos):

📦 CATALOG (11 cambios):
   - ValidateComboSubstitution: price_delta y max_price_delta como int
   - MenuItem.finalPrice(): tipo retorno float → int, base_price como int
   - MenuItem.subtotal componente: (int) round(base_price * quantity)
   - Product: baseAmount, price de lista, base_price como int
   - Product.precioConIva(): (int) round(base_price * (1 + rate/100))
   - MenuItemProduct.subtotal(): tipo retorno float → int

📦 ORDERS (7 cambios):
   - OrderResource: subtotal, tax_amount, discount_amount, total como int
   - OrderItemResource: unit_price, subtotal, price_adjustment como int

NO TOCADOS (float legítimo, 2 casos):
   - Product:216 → tax_rate (tasa de impuesto)
   - OrderItemController:80,84 → tax_rate (tasa de impuesto)

RESULTADO:
✅ 1021 tests passing, 0 failed
✅ 3050 assertions
✅ Consistencia con ADR-018 (CLP integer end-to-end)

PRÓXIMO PASO: Paso 3 (Cashier ~20 casos)

+++ app/Modules/Catalog/Application/UseCases/ValidateComboSubstitution.php
@@ -104,10 +104,10 @@ class ValidateComboSubstitution
         }
 
         // 9. Calcular delta de precio
-        $unitPriceDelta = max(0, (float) $replacementProduct->base_price - (float) $originalProduct->base_price);
+        $unitPriceDelta = max(0, (int) $replacementProduct->base_price - (int) $originalProduct->base_price);  // ADR-018
 
         // 10. Validar max_price_delta si está definido
-        if ($matchedRule->max_price_delta !== null && $unitPriceDelta > (float) $matchedRule->max_price_delta) {
+        if ($matchedRule->max_price_delta !== null && $unitPriceDelta > (int) $matchedRule->max_price_delta) {  // ADR-018
             return SubstitutionValidationResult::denied(
                 'exceeds_max_price_delta',
                 "El recargo unitario ({$unitPriceDelta}) excede el máximo permitido ({$matchedRule->max_price_delta})."
